[Topi] Breakdown topi.cc into smaller files
